+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.0.dev0 (unreleased)
 ---------------------
 
+- Remove b'' from values in py3 csv export
+  [agitator]
+
 - Add contact information to export
   [jensens, agitator]